+++ piargus/__init__.py
@@ -3,13 +3,13 @@ from piargus.batchwriter import BatchWriter
 from piargus.constants import *
 from piargus.inputdata import InputData, MetaData, MicroData, TableData, CodeList
 from piargus.inputdata.hierarchy import Hierarchy, FlatHierarchy, TreeHierarchy, \
-    TreeHierarchyNode, LevelHierarchy
+    TreeHierarchyNode, Node, LevelHierarchy
 from piargus.job import Job, JobSetupError
 from piargus.table import Table, Apriori, TreeRecode
 from piargus.table.safetyrule import *
 from piargus.tauargus import TauArgus
 
-__version__ = "0.4.1"
+__version__ = "1.0.0"
 
 __all__ = [
     "Apriori",
@@ -29,9 +29,11 @@ __all__ = [
     "FlatHierarchy",
     "TreeHierarchy",
     "TreeHierarchyNode",
+    "Node",
     "LevelHierarchy",
 
     # Safety rules
+    "SafetyRule",
     "dominance_rule",
     "percent_rule",
     "frequency_rule",