+++ piargus/inputdata/hierarchy/__init__.py
@@ -1,6 +1,7 @@
 from .levelhierarchy import LevelHierarchy
 from .flathierarchy import FlatHierarchy
-from .treehierarchy import TreeHierarchy, TreeHierarchyNode
+from .treehierarchy import TreeHierarchy, TreeHierarchyNode, Node
 from .hierarchy import Hierarchy
 
-__all__ = ["Hierarchy", "TreeHierarchy", "LevelHierarchy", "FlatHierarchy", "TreeHierarchyNode"]
+__all__ = ["Hierarchy", "TreeHierarchy", "LevelHierarchy", "FlatHierarchy", "TreeHierarchyNode",
+           "Node"]