+++ piargus/inputdata/hierarchy/flathierarchy.py
@@ -8,3 +8,6 @@ class FlatHierarchy(Hierarchy):
 
     def __init__(self, *, total_code=DEFAULT_TOTAL_CODE):
         self.total_code = total_code
+
+    def __repr__(self):
+        return f"{self.__class__.__name__}(total_code={self.total_code})"